# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  company: string;
  phone?: string;
  /** Approximate number of employees to train */
  teamSize?: string;
  /** Primary training requirement (e.g., Data Science, Leadership) */
  trainingArea: string;
  message?: string;
}

export interface Lead {
  id: string;
  name: string;
  email: string;
  company: string;
  phone?: string;
  teamSize?: string;
  trainingArea: string;
  message?: string;
  createdAt: string;
}

export interface LeadResponse {
  message: string;
  lead: Lead;
}

export interface Program {
  id: string;
  title: string;
  category: string;
  duration: string;
  level: string;
  description: string;
  outcomes: string[];
  modules: string[];
}

export interface Solution {
  id: string;
  title: string;
  tagline: string;
  description: string;
  highlights: string[];
  /** Hex color or palette key for visual accent */
  accent: string;
}

export interface Stats {
  professionalsTrained: number;
  batchesLaunched: number;
  satisfactionScore: number;
  enterprisePartners: number;
  countriesReached: number;
  learningHours: number;
}

export interface Testimonial {
  id: string;
  quote: string;
  author: string;
  role: string;
  company: string;
}

export interface Faq {
  id: string;
  question: string;
  answer: string;
}