# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit an enterprise training lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listLeads
      tags: [leads]
      summary: List recent leads (admin / demo only)
      responses:
        "200":
          description: Recent leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LeadInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        company:
          type: string
          minLength: 1
        phone:
          type: string
        teamSize:
          type: string
          description: Approximate number of employees to train
        trainingArea:
          type: string
          description: Primary training requirement (e.g., Data Science, Leadership)
        message:
          type: string
      required:
        - name
        - email
        - company
        - trainingArea
    Lead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        company:
          type: string
        phone:
          type: string
        teamSize:
          type: string
        trainingArea:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - company
        - trainingArea
        - createdAt
    LeadResponse:
      type: object
      properties:
        message:
          type: string
        lead:
          $ref: "#/components/schemas/Lead"
      required:
        - message
        - lead
...