# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: string;
  name: string;
  email: string;
  company: string;
  phone?: string;
  teamSize?: string;
  trainingArea: string;
  message?: string;
  createdAt: Date;
}