# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  company: string;
  phone?: string;
  /** Approximate number of employees to train */
  teamSize?: string;
  /** Primary training requirement (e.g., Data Science, Leadership) */
  trainingArea: string;
  message?: string;
}