# lib/api-zod/src/generated/types/leadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from "./lead";

export interface LeadResponse {
  message: string;
  lead: Lead;
}